Use the correct hostname when sphinx is running on util slice
Currently, when you set sphinx to run on a util slice, the address
in sphinx.yml is hardcoded to point to localhost. Obviously, this
wont work. Altered the config and recipe to use the sphinx util
slice hostname as the address, and still maintain localhost for the
util slice sphinx.yml